+++ repository/src/main/java/com/epam/esm/entity/TagMapper.java
@@ -10,7 +10,7 @@ public class TagMapper implements RowMapper<Tag> {
   @Override
   public Tag mapRow(ResultSet rs, int rowNum) throws SQLException {
     Tag tag = new Tag();
-    tag.setId(rs.getInt("id"));
+    tag.setId(rs.getLong("id"));
     tag.setName(rs.getString("name"));
     return tag;
   }
